### 2024-10-28 #7415 @editorialbot
[PRE REVIEW]: SynxFlow: A GPU-accelerated Python package for multi-hazard simulations
**Submitting author:** @xiaxilin (<a href="http://orcid.org/0000-0002-5784-9211">Xilin Xia</a>)
**Repository:** https://github.com/SynxFlow/SynxFlow
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
