### 2024-10-28 #7415 @kthyng
@xiaxilin
* I see the tutorials/examples in your docs can be run by the users, and these can act as sort of tests for users to verify code usage. Unit tests to test the code would be good to have as well; this may also come up in review.
* Please check that your inline vs. parenthetical references are correct (for example line 51). You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html.
* Also include space before your references for example lines 47 and 48.
* We have a backlog of submissions so I will add this to our waitlist. Thanks for your patience.
