+++ joss/paper.md
@@ -47,10 +47,10 @@ The SynxFlow project is our attempt to achieve this vision.
 The software is a Python package with the following two parts:
 
 ## Solvers
-There are three modules, flood, landslide and debris, solving the shallow water type equations using CUDA-enabled multiple GPUs, to simulate flood inundation, landslides runout and debris flows respectively. The solvers can predict the depth and velocities of these hazards. The flood solver, inherited from Pypims[@Xia2023b], is a Python-bound version of the open-source code hipims-cuda[@HEMLab2020]. Since the start of the SynxFlow project, the landslide solver of 'hipims-cuda' has also been Python-bound and added to SynxFlow, and a debris flow solver has been implemented. These solvers use Godunov-type shock-capturing finite volume methods, which are documented in [@Xia2017; @Xia2018d; @Xia2023]. The numerical methods have been extensively tested in both theoretical and real-world scenarios, demonstrating robustness and efficiency for practical applications.
+There are three modules, flood, landslide and debris, solving the shallow water type equations using CUDA-enabled multiple GPUs, to simulate flood inundation, landslides runout and debris flows respectively. The solvers can predict the depth and velocities of these hazards. The flood solver, inherited from Pypims [@Xia2023b], is a Python-bound version of the open-source code hipims-cuda [@HEMLab2020]. Since the start of the SynxFlow project, the landslide solver of hipims-cuda has also been Python-bound and added to SynxFlow, and a debris flow solver has been implemented. These solvers use Godunov-type shock-capturing finite volume methods, which are documented in @Xia2017, @Xia2018d and @Xia2023. The numerical methods have been extensively tested in both theoretical and real-world scenarios, demonstrating robustness and efficiency for practical applications.
 
 ## IO (Input and Output)
-The IO module comprises of several classes and functions that generates the inputs to the solvers and process their outputs. The IO functionality for the flood solver is based on the open-source 'hipims-io'[@Ming2023] package, which has been extended to support the landslide and debris solvers. To set up a simulation, the users need to define an object using the 'InputModel' class, using common data formats including numpy arrays and pandas dataframes among others. An 'OutputModel' type object can be used to perform a variety of visualisation tasks including plotting maps, hydrographs and creating animations. To assist users, the IO module also provide functionalities to process geospatial data such as rasters and shapefiles. The IO module effectively integrates the solvers into data science workflows, streamlining hazard risk assessment tasks, and also enables coupling between solvers to facilitate multi-hazard simulations.
+The IO module comprises of several classes and functions that generates the inputs to the solvers and process their outputs. The IO functionality for the flood solver is based on the open-source hipims-io [@Ming2023] package, which has been extended to support the landslide and debris solvers. To set up a simulation, the users need to define an object using the 'InputModel' class, using common data formats including numpy arrays and pandas dataframes among others. An 'OutputModel' type object can be used to perform a variety of visualisation tasks including plotting maps, hydrographs and creating animations. To assist users, the IO module also provide functionalities to process geospatial data such as rasters and shapefiles. The IO module effectively integrates the solvers into data science workflows, streamlining hazard risk assessment tasks, and also enables coupling between solvers to facilitate multi-hazard simulations.
 
 Compared with existing software, SynxFlow is unique for its combination of open-source accessibility, multi-GPU acceleration, multi-hazard simulation capabilities, and a Python-based user interface.
 
