### 2024-10-28 #7415 @xiaxilin
@kthyng

Thank you for promptly looking into this submission. I have updated the paper to address your comments about the references.
