### 2024-11-05 #7415 @christopheredsall
> @christopheredsall,  would you review this submission for JOSS?

Sure, I can give that a go.
