### 2024-12-11 #7586 @editorialbot
[REVIEW]: SynxFlow: A GPU-accelerated Python package for multi-hazard simulations
**Submitting author:** @xiaxilin (<a href="http://orcid.org/0000-0002-5784-9211">Xilin Xia</a>)
**Repository:** https://github.com/SynxFlow/SynxFlow
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
