+++ synxflow/CMakeLists.txt
@@ -25,7 +25,7 @@ include(FetchContent)
 FetchContent_Declare(
   pybind11_sources
   GIT_REPOSITORY https://github.com/pybind/pybind11.git
-  GIT_TAG        v2.5
+  GIT_TAG        v2.12.0
 )
 
 FetchContent_GetProperties(pybind11_sources)
