+++ synxflow/apps/cudaDebrisFlowSolverPybind/debris_solver_pybind.cc
@@ -19,7 +19,28 @@
 namespace py = pybind11;
 
 PYBIND11_MODULE(debris, m) {
-  m.doc() = "core engine for debris flow model"; // optional module docstring
+  m.doc() = R"pbdoc(
+        Debris Flow Simulation Engine
+        ==============================
 
-  m.def("run", &run, "debris flow simulation with single gpu");
+        This module provides the core functionality for running a debris flow simulation.
+    )pbdoc"; // optional module docstring
+
+  m.def("run", &run, R"pbdoc(
+        run(work_dir)
+
+        Runs the debris flow simulation using the specified working directory.
+
+        Parameters:
+            work_dir (str): A path to the working directory where all necessary input
+                files (such as configuration or data files) are located. This directory
+                is also used to store the simulation outputs.
+
+        Returns:
+            int: A status code indicating the result of the simulation. A value of 0 typically
+                means the simulation completed successfully, while a non-zero value indicates an error.
+
+        Note:
+            Ensure that the working directory is correctly set up with the required input data.
+    )pbdoc");
 }
