+++ synxflow/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.0.5.0'
+__version__ = '1.0.1'
