+++ synxflow/apps/cudaLandslideSolverPybind/landslide_solver_pybind.cc
@@ -19,7 +19,27 @@
 namespace py = pybind11;
 
 PYBIND11_MODULE(landslide, m) {
-  m.doc() = "core engine for landslide runout model"; // optional module docstring
+    m.doc() = R"pbdoc(
+        Landslide Runout Simulation Engine
+        ====================================
 
-  m.def("run", &run, "landslide runout simulation with single gpu");
+        This module provides the core functionality for running a landslide runout simulation.
+    )pbdoc";
+
+    m.def("run", &run, R"pbdoc(
+        run(work_dir)
+
+        Executes a landslide runout simulation using a single GPU.
+
+        Parameters:
+            work_dir (str): The path to the working directory that contains all the necessary
+                input data and configuration files required for the simulation.
+
+        Returns:
+            int: A status code indicating the result of the simulation. A value of 0 typically
+                means that the simulation completed successfully, while any non-zero value indicates
+                that an error occurred.
+
+    )pbdoc");
 }
+
