+++ setup.py
@@ -80,9 +80,9 @@ setup(
     author='Xilin Xia',
     author_email='x.xia.1@bham.ac.uk',
     packages=find_packages(),
-    ext_modules=[CMakeExtension('synxflow/apps/cudaFloodSolversPybind/cuda_debris_flow_solver_pybind'),
-                 CMakeExtension('synxflow/apps/cudaDebrisFlowSolverPybind/cuda_flood_solvers_pybind'),
-                 CMakeExtension('synxflow/apps/cudaLandslideSolverPybind/cuda_landslide_solver_pybind')],
+    ext_modules=[CMakeExtension('synxflow.apps.cudaFloodSolversPybind.cuda_debris_flow_solver_pybind'),
+                 CMakeExtension('synxflow.apps.cudaDebrisFlowSolverPybind.cuda_flood_solvers_pybind'),
+                 CMakeExtension('synxflow.apps.cudaLandslideSolverPybind.cuda_landslide_solver_pybind')],
     long_description=open("README.md").read(),
     long_description_content_type='text/markdown',
     description='Simulating Mass Movement and Flood Inundation for Multi-hazard Risk Assessment',
