+++ docs/source/Tutorials/LLM.md
@@ -1,0 +1,236 @@
+# Tutorial: Using LLMs to Build SynxFlow Workflow
+
+Authour: [Xilin Xia](https://www.linkedin.com/in/xilin-xia-8787a4199/?originalSubdomain=uk)
+
+This guide explains how you can leverage Large Large Language Model as your assistant for using SynxFlow. I will use OpenAI's ChatGPT o3-mini as an example, which is good at coding tasks and can access information on the internet. But you can also use other LLMs with similar capabilities.
+
+---
+
+## Step 1: Browsing the Documentation (The First Prompt)
+
+This is to ensure that ChatGPT can learn the latest information about the API usage of SynxFlow.
+
+**First Prompt:**
+
+> "Please use your browsing capability to visit and read the following SynxFlow documentation pages and tutorials:  
+> 1. InputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html)  
+> 2. Boundary API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html](https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html)  
+> 3. Rainfall API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html](https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html)  
+> 4. OutputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html)  
+> 5. Raster API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Raster.html](https://synxflow.readthedocs.io/en/latest/Modules/Raster.html) 
+> 6. Tutorials: (for practical examples of using SynxFlow) [https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html](https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html)
+> 7. Installation: (for steps of install the model) [https://synxflow.readthedocs.io/en/latest/quickstart.html](https://synxflow.readthedocs.io/en/latest/quickstart.html)
+
+> After reviewing, summarise key functionalities of SynxFlow."
+
+This first prompt instructs ChatGPT to browse and understand the key components of SynxFlow. By reviewing these pages, ChatGPT gathers detailed information about:
+
+- **InputModel:** How to define inputs (DEM, rainfall, boundaries, etc.)  
+- **Boundary:** How to set up different boundary conditions (including the "fall" type)  
+- **Rainfall:** How to handle time series and spatial distribution of rainfall  
+- **OutputModel:** How to read and analyze simulation outputs  
+- **Raster:** How to handle gridded data and visualize them
+- **Tutorials:** Practical, step-by-step examples for running flood simulations
+
+This step is crucial because it ensures that any examples or code generated are informed by the latest SynxFlow documentation and best practices.
+
+---
+
+## Step 2: Setting Up a Simulation Case with ChatGPT Assistance
+
+Once the documentation has been reviewed, you can ask ChatGPT to generate a code example for setting up a simulation. For instance, you might ask:
+
+> “Please show me a sample code that creates, runs, and visualizes a flood simulation using SynxFlow’s sample data.”
+
+ChatGPT will then provide a complete script that:
+- Loads sample data (DEM, rain mask, landcover, rain source)  
+- Initializes an **InputModel** with proper initial conditions, boundary conditions, and rainfall inputs  
+- Sets up landcover-based parameters (such as Manning’s n)  
[…]
