+++ docs/source/Tutorials/LLM.md
@@ -1,4 +1,4 @@
-# Tutorial: Using LLMs to Build SynxFlow Workflow
+# Using LLMs to Build SynxFlow Workflow
 
 Authour: [Xilin Xia](https://www.linkedin.com/in/xilin-xia-8787a4199/?originalSubdomain=uk)
 
@@ -12,16 +12,19 @@ This is to ensure that ChatGPT can learn the latest information about the API us
 
 **First Prompt:**
 
-> "Please use your browsing capability to visit and read the following SynxFlow documentation pages and tutorials:  
-> 1. InputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html)  
-> 2. Boundary API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html](https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html)  
-> 3. Rainfall API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html](https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html)  
-> 4. OutputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html)  
-> 5. Raster API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Raster.html](https://synxflow.readthedocs.io/en/latest/Modules/Raster.html) 
-> 6. Tutorials: (for practical examples of using SynxFlow) [https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html](https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html)
-> 7. Installation: (for steps of install the model) [https://synxflow.readthedocs.io/en/latest/quickstart.html](https://synxflow.readthedocs.io/en/latest/quickstart.html)
+> *“Please use your browsing capability to visit and read the following SynxFlow documentation pages and tutorials:*
 
-> After reviewing, summarise key functionalities of SynxFlow."
+> *1. InputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html)*  
+> *2. Boundary API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html](https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html)*  
+> *3. Rainfall API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html](https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html)*  
+> *4. OutputModel API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html](https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html)*  
+> *5. Raster API Documentation: [https://synxflow.readthedocs.io/en/latest/Modules/Raster.html](https://synxflow.readthedocs.io/en/latest/Modules/Raster.html)*  
+> *6. Tutorials: (for practical examples of using SynxFlow) [https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html](https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html)*  
+> *7. Installation: (for steps of install the model) [https://synxflow.readthedocs.io/en/latest/quickstart.html](https://synxflow.readthedocs.io/en/latest/quickstart.html)*  
+
+> *After reviewing, summarise key functionalities of SynxFlow.”*
+
+
 
[…]
