+++ docs/source/Tutorials/LLM.rst
@@ -1,0 +1,244 @@
+==============================================
+Using LLMs to Build SynxFlow Workflow
+==============================================
+
+This guide explains how you can leverage a Large Language Model (LLM) as your assistant when using SynxFlow. In this example, OpenAI's ChatGPT o3-mini is used because it is effective at coding tasks and can access internet information. (Other LLMs with similar capabilities can also be used.)
+
+------------------------------------------------------------
+Step 1: Browsing the Documentation (The First Prompt)
+------------------------------------------------------------
+
+This step ensures that the LLM learns the latest API usage and best practices of SynxFlow by reviewing its online documentation.
+
+**First Prompt:**
+
+.. admonition:: Example Prompt
+   :class: tip
+
+   “Please use your browsing capability to visit and read the following SynxFlow documentation pages and tutorials:
+
+   1. InputModel API Documentation: `https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html <https://synxflow.readthedocs.io/en/latest/Modules/InputModel.html>`_
+   2. Boundary API Documentation: `https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html <https://synxflow.readthedocs.io/en/latest/Modules/Boundary.html>`_
+   3. Rainfall API Documentation: `https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html <https://synxflow.readthedocs.io/en/latest/Modules/Rainfall.html>`_
+   4. OutputModel API Documentation: `https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html <https://synxflow.readthedocs.io/en/latest/Modules/OutputModel.html>`_
+   5. Raster API Documentation: `https://synxflow.readthedocs.io/en/latest/Modules/Raster.html <https://synxflow.readthedocs.io/en/latest/Modules/Raster.html>`_
+   6. Tutorials (practical examples of using SynxFlow): `https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html <https://synxflow.readthedocs.io/en/latest/Tutorials/flood.html>`_
+   7. Installation (steps to install SynxFlow): `https://synxflow.readthedocs.io/en/latest/quickstart.html <https://synxflow.readthedocs.io/en/latest/quickstart.html>`_
+
+   After reviewing, summarise the key functionalities of SynxFlow.”
+
+By reviewing these pages, the LLM gathers detailed information about:
+
+- **InputModel:** How to define inputs (e.g., DEM, rainfall, boundaries).
+- **Boundary:** How to set up different boundary conditions.
+- **Rainfall:** How to handle time series and spatial rainfall data.
+- **OutputModel:** How to read and analyze simulation outputs.
+- **Raster:** How to process gridded data and visualize them.
+- **Tutorials:** Step-by-step examples for running flood simulations.
+- **Installtation:** Steps to install SynxFlow.
+
+------------------------------------------------------------
+Step 2: Setting Up a Simulation Case with LLM Assistance
+------------------------------------------------------------
+
+Once the documentation has been reviewed, you can ask the LLM to generate a sample code example for setting up a simulation. For example:
+
+.. admonition:: Example Prompt
+   :class: tip
+
+   “Please show me a sample code that creates, runs, and visualizes a flood simulation using SynxFlow’s sample data.”
+
+The LLM will generate a script that:
+- Loads sample data (DEM, rain mask, landcover, rain source)
+- Initializes an **InputModel** with the proper initial conditions, boundary conditions, and rainfall inputs
+- Sets up landcover-based parameters (e.g., Manning’s n)
+- Writes input files and runs the simulation using **flood.run()**
+- Reads simulation outputs with **OutputModel** and visualizes them (using **Raster.mapshow()**)
+
+------------------------------------------------------------
[…]
