+++ docs/source/Modules/engines.rst
@@ -2,35 +2,42 @@
 Simulation Engines
 ==========================================
 
-The SynxFlow package includes multiple simulation engines implemented as C extensions.
-Each engine is exposed as a Python class with a common interface. The following sections describe
+The SynxFlow package includes multiple simulation engine modules implemented as CUDA extensions.
+Each engine is exposed as a Python module with a common interface. The following sections describe
 the API for each simulation engine:
 
 - **flood**: Flood simulation engine.
 - **landslide**: Landslide runout engine.
 - **debris**: Debris flow simulation engine.
 
-.. py:class:: flood(work_dir)
-
-   :param work_dir: str
-      The path to the working directory that contains all necessary input files (e.g. digital elevation models, rainfall data)
-      and where simulation outputs will be stored.
+.. py:module:: flood
 
    **Overview:**
 
-   The *FloodEngine* class provides the core functionality for simulating flood dynamics using GPU acceleration.
-   It performs the simulation based on the provided input data and returns a status code indicating success or failure.
+   The *flood* module provides the core functionality for simulating flood dynamics using GPU acceleration.
 
-   **Methods:**
+   **Function: run(work_dir)**
 
-   .. py:method:: run()
-      
+   .. py:function:: run(work_dir)
+
       Executes the flood simulation.
 
-      **Returns:**
+      :param work_dir: str
+         The path to the working directory that contains all necessary input files (e.g. digital elevation models, rainfall data)
+         and where simulation outputs will be stored.
+      :returns: int
+         A status code where a return value of 0 indicates that the simulation completed successfully,
+         and any non-zero value indicates an error.
 
-         int  
-         A status code where a return value of 0 indicates the simulation completed successfully,
+   .. py:function:: run_mgpus(work_dir)
+
+      Executes the flood simulation using multiple GPUs.
+
+      :param work_dir: str
+         The path to the working directory that contains all necessary input files (e.g. digital elevation models, rainfall data)
+         and where simulation outputs will be stored.
+      :returns: int
+         A status code where a return value of 0 indicates that the simulation completed successfully,
          and any non-zero value indicates an error.
 
    **Example:**
@@ -39,35 +46,29 @@ the API for each simulation engine:
 
       from synxflow import flood
 
-      # Create an instance of the FloodEngine with the specified working directory
-      engine = flood.FloodEngine("/path/to/flood/work_dir")
-      status = engine.run()
+      status = flood.run("/path/to/flood/work_dir")
       if status == 0:
           print("Flood simulation completed successfully!")
       else:
           print("Flood simulation encountered an error.")
 
 
-.. py:class:: landslide(work_dir)
-
-   :param work_dir: str
-      The path to the working directory containing the necessary input data (e.g. terrain models, material properties)
-      for the landslide simulation, and where outputs will be written.
+.. py:module:: landslide
 
    **Overview:**
 
-   The *LandslideEngine* class implements the core engine for simulating landslide runout events.
-   It utilizes GPU acceleration to compute the dynamics of landslide movement based on input configurations.
+   The *landslide* module provides the core functionality for simulating landslide runout events using GPU acceleration.
 
-   **Methods:**
+   **Function: run(work_dir)**
 
-   .. py:method:: run()
-      
+   .. py:function:: run(work_dir)
+
       Starts the landslide runout simulation.
 
-      **Returns:**
-
-         int  
+      :param work_dir: str
+         The path to the working directory containing the necessary input data (e.g. terrain models, material properties)
+         for the landslide simulation, and where outputs will be written.
+      :returns: int
          A status code where 0 indicates success and a non-zero value indicates an error.
 
    **Example:**
[…]
