### 2025-02-03 a7a2706
fix small issues with doc
