### 2025-02-03 #7586 @xiaxilin
@PabloIbannez thank you for your comments, which are really useful. I have updated the code and documentation so that

1. It runs on Colab again, this was due to recently Colab upgraded to Python 3.11. So I have built a python 3.11 binary
2. The documentation now clearly says landslide and debris only runs on single GPU.
3. API reference for flood, landslide and debris has been added.

I also added a section about how to use LLM to generate script for running SynxFlow.
