### 2025-02-08 #7586 @PabloIbannez
Thank you for your prompt response and for improving both the tests and the documentation. I especially appreciate the addition of LLM-generated scripts, which is an original and very useful approach.

That said, I noticed that there are no formal automated tests. While I personally consider the published references and prior applications of the software as sufficient validation, this is ultimately for the JOSS editor to decide (@AnjaliSandip). Additionally, I suggest expanding the discussion on the state of the field, as the current references are exclusively to the authors' previous work. A broader comparison with other commonly used software would strengthen the paper.

That being said, I want to emphasize that SynxFlow appears to be an excellent project. Congratulations! It is great to see that the library is already gaining traction. I recommend publication once the paper and repository are slightly improved, particularly regarding the state of the field section. Lastly, a careful second proofreading would be beneficial, as there are a few minor language issues.
