+++ synxflow/__init__.py
@@ -10,23 +10,10 @@ Created on Wed Apr  1 14:56:15 2020
 """
 
 from .version import __version__
-
-# Defer imports to avoid circular dependencies
-def _load_modules():
-    from . import flood, IO
-    return flood, IO
+from . import flood, IO
 
 __all__ = [
-    '__version__',
-    'flood',  # Will be populated by __getattr__
-    'IO',     # Will be populated by __getattr__
+  '__version__',
+  'flood',
+  'IO',
 ]
-
-# Lazy loading of modules
-def __getattr__(name):
-    if name in ['flood', 'IO']:
-        flood, IO = _load_modules()
-        globals()['flood'] = flood
-        globals()['IO'] = IO
-        return globals()[name]
-    raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
