### 2025-02-12 32fa156
Added Docker Support 1, Added a new Dockerfile for containerizing Synxflow 2, Added a readme file with instructions on how to run the docker container.
