### 2025-02-12 #7586 @thivinanandh
@AnjaliSandip - Sorry for the delay
@xiaxilin - Thank you for your patience. I have completed my review on SynxFlow and it seems to be an excellent library for hazard flow simulation in Python. After reviewing the package, I have some small recommendations that could make this library even better.

I have raised an issue in the target repository [Here](https://github.com/SynxFlow/SynxFlow/issues/43). The summary of the recommendations are

1. **Paper**: Expand little more on existing flood simulation softwares that are available and clarify how this library compares to current packages.
2. **Documentation**: Add some documentation on file types and data files used in example files.

3. **Automated Test Cases**: Implement unit testing for the framework.

4. **Docker Containerization**: Since the framework is GPU dependent, it will be much better to also have a Docker image for this container, which makes the users to run it on any environment

I have also added PR's for Docker ( [link](https://github.com/SynxFlow/SynxFlow/pull/44) ) and unit tests ( [link](https://github.com/SynxFlow/SynxFlow/pull/45) ), which you can go though..
