+++ tests/unit_tests/test_input.py
@@ -1,0 +1,154 @@
+import unittest
+import numpy as np
+import os
+import shutil
+from synxflow.IO import InputModel, Raster
+from synxflow.IO.demo_functions import get_sample_data
+
+
+class TestInputModelWithSampleData(unittest.TestCase):
+
+    def setUp(self):
+        """Set up test fixtures, load sample data and create temp folders."""
+        print("Setting up test with sample data...")
+        # Get sample data paths and dictionary using the function from the tutorial
+        self.dem_file, self.sample_data, self.data_path = get_sample_data() # [cite: 447, 597]
+
+        # Load the actual sample DEM
+        print(f"Loading sample DEM: {self.dem_file}")
+        self.dem_raster = Raster(self.dem_file) # [cite: 606]
+        self.rows, self.cols = self.dem_raster.shape
+        self.header = self.dem_raster.header
+
+        # Load sample rain mask (needed for rainfall test)
+        self.rain_mask_file = os.path.join(self.data_path, 'rain_mask.gz')
+        print(f"Loading sample rain mask: {self.rain_mask_file}")
+        self.rain_mask_raster = Raster(self.rain_mask_file) # [cite: 608]
+
+        # Load sample landcover (needed for landcover-based param test if added)
+        self.landcover_file = os.path.join(self.data_path, 'landcover.gz')
+        print(f"Loading sample landcover: {self.landcover_file}")
+        self.landcover_raster = Raster(self.landcover_file) # [cite: 611]
+
+        # Create a temporary directory for test outputs
+        self.test_dir = "temp_test_input_model_sample"
+        if os.path.exists(self.test_dir):
+            shutil.rmtree(self.test_dir)
+        os.makedirs(self.test_dir)
+        print(f"Temporary directory created: {self.test_dir}")
+
+    def tearDown(self):
+        """Tear down test fixtures, remove temporary directory."""
+        print("Tearing down test...")
+        if os.path.exists(self.test_dir):
+            shutil.rmtree(self.test_dir)
+            print(f"Temporary directory removed: {self.test_dir}")
+
+    def test_initialization_with_sample_raster(self):
+        """Test InputModel initialization using the sample Raster object."""
+        model = InputModel(dem_data=self.dem_raster, case_folder=self.test_dir)
+        self.assertIsInstance(model.DEM, Raster)
+        self.assertEqual(model.shape, self.dem_raster.shape)
+        self.assertEqual(model.header['ncols'], self.header['ncols'])
+        self.assertEqual(model.num_of_sections, 1)
+        self.assertEqual(model.get_case_folder(), self.test_dir)
+        # Check default attributes
+        self.assertEqual(model.attributes['manning'], 0.035) # Default Manning's n
+        self.assertTrue(hasattr(model, 'Boundary'))
+        self.assertTrue(hasattr(model, 'Rainfall'))
+
+    def test_set_boundary_condition_with_sample(self):
+        """Test setting boundary conditions using sample data."""
+        model = InputModel(dem_data=self.dem_raster, case_folder=self.test_dir)
+        # Use boundary_condition data from the loaded sample_data dict [cite: 447, 621]
+        boundary_list_sample = self.sample_data['boundary_condition']
+        model.set_boundary_condition(boundary_list=boundary_list_sample, outline_boundary='fall') # [cite: 621] Matches tutorial default
+
+        self.assertEqual(model.Boundary.num_of_bound, 3) # 2 defined + 1 outline
+        self.assertEqual(model.Boundary.data_table.type[0], 'fall') # Check outline type
[…]
