+++ tests/unit_tests/test_rainfall.py
@@ -1,0 +1,136 @@
+import unittest
+import numpy as np
+import os
+import shutil
+import sys # Keep for skipIf
+
+from synxflow.IO.Rainfall import Rainfall
+from synxflow.IO.Raster import Raster
+from synxflow.IO.demo_functions import get_sample_data
+
+class TestRainfallRefocused(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        """Load sample data paths and base DEM once."""
+        print("Setting up class: Loading sample data paths...")
+        try:
+            # Get paths to sample data included in the package
+            cls.dem_file, cls.demo_data_dict, cls.data_path = get_sample_data(case_type='flood')
+            cls.sample_mask_path = os.path.join(cls.data_path, 'rain_mask.gz')
+            cls.sample_source_path = os.path.join(cls.data_path, 'rain_source.csv')
+            # Load the base DEM raster for context
+            cls.base_dem_raster = Raster(cls.dem_file)
+            cls.sample_data_available = True
+        except Exception as e:
+            print(f"Warning: Could not load sample data via get_sample_data: {e}")
+            cls.sample_data_available = False
+            # Define dummy raster if needed for tests that don't rely on loading sample
+            cls.base_dem_raster = Raster(array=np.ones((4,5)), header={'ncols': 5, 'nrows': 4,'xllcorner': 0, 'yllcorner': 0,'cellsize': 10, 'NODATA_value': -9999.0})
+
+
+    def setUp(self):
+        """Set up test fixtures before each test method (minimal setup needed now)."""
+        # Create a temporary directory for potential temporary file needs (if any test needed it)
+        self.test_dir = f"temp_test_rainfall_refocused_{self.id().split('.')[-1]}"
+        os.makedirs(self.test_dir, exist_ok=True)
+
+    def tearDown(self):
+        """Tear down test fixtures after each test method."""
+        if os.path.exists(self.test_dir):
+            shutil.rmtree(self.test_dir)
+
+    @unittest.skipUnless(lambda: hasattr(TestRainfallRefocused, 'sample_data_available') and TestRainfallRefocused.sample_data_available,
+                         "Sample data not loaded, skipping sample load test")
+    def test_load_sample_rainfall_data(self):
+        """Test successful loading and processing of sample rainfall data from package."""
+        # Uses class variables set up in setUpClass
+        rainfall = Rainfall(rain_mask=self.sample_mask_path,
+                            rain_source=self.sample_source_path,
+                            source_sep=',', # Specify separator for csv
+                            dem_ras=self.base_dem_raster)
+
+        # Verify essential attributes exist and seem reasonable
+        self.assertIsNotNone(rainfall.mask_header, "Mask header not initialized")
+        self.assertTrue(hasattr(rainfall, 'time_s'), "Time series attribute missing")
+        self.assertTrue(hasattr(rainfall, 'rain_rate'), "Rain rate attribute missing")
+        self.assertTrue(hasattr(rainfall, 'subs_in'), "Subscripts attribute missing (from dem_ras)")
+        self.assertTrue(hasattr(rainfall, 'attrs'), "Attributes dict missing")
+        self.assertIsInstance(rainfall.attrs, dict)
+        self.assertTrue(len(rainfall.time_s) > 0, "Time series is empty")
+        self.assertTrue(rainfall.rain_rate.shape[0] == len(rainfall.time_s), "Rain rate rows != time series length")
+        self.assertTrue(rainfall.rain_rate.shape[1] > 0, "Rain rate has no sources")
+
+
+    def test_init_invalid_mask_path(self):
+        """Test error handling for non-existent mask file path."""
+        non_existent_mask = os.path.join(self.test_dir, "non_existent_mask.gz")
[…]
