+++ synxflow/IO/demo_functions.py
@@ -95,9 +95,9 @@ def get_sample_data(case_type = 'flood'):
                                     [72, 1170]]),
                 'type': 'open',
                 'h': [[0, 10], [60, 10]]}],
-            'rain_source': np.array([[0, 2.77777778e-06],
-                                    [1800, 2.77777778e-05],
-                                    [3600, 0]]),
+            'rain_source': np.array([[0, 2.77777778e-06, 0, 0, 0],
+                                    [1800, 2.77777778e-05, 0, 0, 0],
+                                    [3600, 0, 0, 0, 0]]),
             'gauges_pos': np.array([[300, 1600],
                                     [400, 1200]])
         }

[…]
