+++ docs/source/Tutorials/data_list.md
@@ -1,0 +1,41 @@
+# Input and output data
+
+This article gives the lists of input and output data of SynxFlow model.
+
+Date: 31/03/2025
+
+## User-Provided Input Files & Data
+
+This table lists inputs typically provided by the user either as files or defined directly in their Python setup script, noting solver specificity.
+
+| User Input Type                 | Example Data/Format                                                                                          | Purpose                                                                      | Notes                                                                                        |
+| :------------------------------ | :------------------------------------------------------------------------------------------------------------------------------------ | :--------------------------------------------------------------------------- | :--------------------------------------------------------------------------------------------------------------------- |
+| **Files Provided by User:** |                                                                                                                                       |                                                                              |                                                                                                                        |
+| DEM File                        | Raster file (`.gz`, `.asc`, `.tif`)                                                                                                   | Defines the topography/elevation                                             | Required for all simulation types.                                                                                   |
+| Rain Mask File (Flood and Debris only)     | Raster file (`.gz`, `.asc`, `.tif`)                                                                                                   | Assigns rainfall source indices to grid cells                                | Primarily for flood and debris models via `set_rainfall`.                                                            |
+| Rain Source File (Flood and Debris only)   | CSV file (`.csv`) with columns: time(s), rate\_src0(m/s), ...                                                                         | Provides time series of rainfall rates for each source                       | Primarily for flood and debris models via `set_rainfall`.                                                            |
+| Land Cover File      | Raster file (`.gz`, `.asc`, `.tif`)                                                                                                   | Assigns land cover type indices for parameter variation                      | Used by `set_landcover` / `set_grid_parameter`. Applicable if parameters vary spatially.                     |
+| **Data Defined in Script:** |                                                                                                                                       |                                                                              |                                                                                                                        |
+| Initial Conditions              | Scalar or NumPy array (e.g., for `h0`, `hU0x`, `hU0y`)                                                                                | Sets initial material depth/mass (`h0`), momentum(velocity times depth) (`hU0x`, `hU0y`)            | Passed to `set_initial_condition()`. Needed for all types. Default values are 0.                                           |
+| Boundary Conditions     | Python List of Dictionaries (defining location, type, time series `h`/`hU` / `C` )                                                           | Defines where/how water enters/leaves the domain                             | Set via `set_boundary_condition()`. |
+| **Grid Parameters:** | Set via `set_grid_parameter()`. Can be Scalar, NumPy array, or Dict (if using Landcover)                               | Defines physical properties across the grid                                  |                                                                                                                        |
+| *Manning's n* (Flood/Debris)    | Scalar/Array/Dict                                                                                                                     | Surface roughness / flow resistance                                          | Flood and debris models only.                                                                                                      |
+| *Infiltration Params* (Flood) | Scalar/Array/Dict (e.g., `hydraulic_conductivity`, `capillary_head`, `water_content_diff`, `sewer_sink`)                          | Parameters for Green-Ampt infiltration and drainage loss                                       | Flood model only.                                                                                             |
[…]
