### 2025-03-31 e525eda
add current status to joss paper
