+++ joss/paper.md
@@ -32,7 +32,11 @@ SynxFlow can dynamically simulate flood inundation, landslides runout and debris
 
 Natural hazards like floods, landslides, and debris flows cause significant damage each year, and due to climate change, the frequency and severity of these events are increasing. This rising threat highlights an urgent need to develop tools that can enhance our understanding of these risks, supporting crucial applications like early warning systems and comprehensive risk assessments.
 
-Effectively assessing these risks requires integrating data from multiple sources—such as climate, geological, hydrological, and environmental data, while also accounting for complex, interacting hazards. For comprehensive risk assessment over a large region, processing vast amounts of data swiftly is essential, often necessitating cloud-based computing or the use of high-performance computing (HPC) systems. These HPC systems, especially when enhanced with Graphics Processing Units (GPUs), allow faster processing of complex simulations. However, most existing software for simulating natural hazards is desktop-based and challenging to integrate into the modern data science ecosystem, making it unsuitable for large-scale, real-time applications and collaborative research efforts. This gap underscores the need for advanced and scalable tools that leverage modern computing infrastructure, enabling seamless integration, real-time data processing, and accessibility for diverse stakeholders.
+Effectively assessing these risks requires integrating data from multiple sources—such as climate, geological, hydrological, and environmental data, while also accounting for complex, interacting hazards. For comprehensive risk assessment over a large region, processing vast amounts of data swiftly is essential, often necessitating cloud-based computing or the use of high-performance computing (HPC) systems. These HPC systems, especially when enhanced with Graphics Processing Units (GPUs), allow faster processing of complex simulations.
+
+However, current modeling tools show significant limitations in meeting the growing demands of multi-hazard risk assessment and in leveraging modern computing technologies. For example, tools like LISFLOOD-FP 8.0 [@Shaw2021] and TRITON [@MoralesHernandez2021] offer GPU acceleration, but their scope is restricted to flood simulations and they lack native Python interfaces — making integration with data science workflows cumbersome. GeoClaw [@clawpack;@BergerLeVeque98] supports multiple hazards and is Python-compatible, yet its single-GPU acceleration makes it inefficient for large-scale, high-resolution applications.
+
+HEC-RAS [@HEC-RAS] is widely used due to its user-friendly graphical interface and support for third-party Python automation. However, it does not support multi-GPU acceleration, cannot simulate landslide runout, and remains proprietary software — limiting both customisation and open development. These constraints highlight a critical gap: existing tools are either too specialised, insufficiently scalable, or poorly integrated with modern data and compute ecosystems.
 
 To this end, we believe modern multi-hazard simulation software should be
 
@@ -50,7 +54,7 @@ The software is a Python package with the following two parts:
 There are three modules, flood, landslide and debris, solving the shallow water type equations using CUDA-enabled multiple GPUs, to simulate flood inundation, landslides runout and debris flows respectively. The solvers can predict the depth and velocities of these hazards. The flood solver, inherited from Pypims [@Xia2023b], is a Python-bound version of the open-source code hipims-cuda [@HEMLab2020]. Since the start of the SynxFlow project, the landslide solver of hipims-cuda has also been Python-bound and added to SynxFlow, and a debris flow solver has been implemented. These solvers use Godunov-type shock-capturing finite volume methods, which are documented in @Xia2017, @Xia2018d and @Xia2023. The numerical methods have been extensively tested in both theoretical and real-world scenarios, demonstrating robustness and efficiency for practical applications.
 
 ## IO (Input and Output)
[…]
