+++ joss/paper.bib
@@ -67,3 +67,51 @@
     volume = {234},
     year = {2018}
 }
+
+@article{Shaw2021,
+  author    = {James Shaw and Georges Kesserwani and Jeffrey Neal and Paul Bates and Mohammad Kazem Sharifian},
+  title     = {{LISFLOOD-FP 8.0: the new discontinuous Galerkin shallow-water solver for multi-core CPUs and GPUs}},
+  journal   = {Geoscientific Model Development},
+  year      = {2021},
+  volume    = {14},
+  pages     = {3577--3602},
+  doi       = {10.5194/gmd-14-3577-2021},
+  url       = {https://gmd.copernicus.org/articles/14/3577/2021/}
+}
+
+@article{MoralesHernandez2021,
+  author    = {Mario Morales-Hernández and Md Bulbul Sharif and Alfred Kalyanapu and S. K. Ghafoor and T. T. Dullo and S. Gangrade and Shih-Chieh Kao and M. R. Norman and K. J. Evans},
+  title     = {{TRITON: A Multi-GPU Open Source 2D Hydrodynamic Flood Model}},
+  journal   = {Environmental Modelling & Software},
+  year      = {2021},
+  volume    = {141},
+  pages     = {105034},
+  doi       = {10.1016/j.envsoft.2021.105034},
+  url       = {https://doi.org/10.1016/j.envsoft.2021.105034}
+}
+
+@misc{HEC-RAS,
+  author    = {{US Army Corps of Engineers}},
+  title     = {{HEC-RAS River Analysis System}},
+  year      = {2025},
+  note      = {Version 2025 Alpha},
+  url       = {https://www.hec.usace.army.mil/software/hec-ras/}
+}
+
+@misc{clawpack,
+    title={Clawpack software},
+    author={{Clawpack Development Team}},
+    url={http://www.clawpack.org},
+    note={Version 5.11.0},
+    doi={https://doi.org/10.5281/zenodo.13376470},
+    year={2024}
+}
+
+@article{BergerLeVeque98,
+    Author = {M. J. Berger and R. J. LeVeque},
+    Journal = {SIAM J. Numer. Anal.},
+    Pages = {2298--2316},
+    Title = {Adaptive Mesh Refinement using Wave-Propagation Algorithms for Hyperbolic Systems},
+    Volume = {35},
+    Year = {1998}
+}
