### 2025-03-31 #7586 @xiaxilin
@AnjaliSandip @PabloIbannez @thivinanandh Thank you for your great feedback and patience! I really appreciate! I have now made the following changes:

1. A review of popular existing software has been added to the paper.
2. Unit test has been added.
3. Dockerisation has been incorporated by merging the pull request by @thivinanandh
4. An explanation for input and output data has been added to the [online documentation](https://synxflow.readthedocs.io/en/latest/Tutorials/data_list.html).

I hope these changes have adequately answered all the questions by you.
