+++ COMMUNITY_GUIDELINES.md
@@ -1,0 +1,213 @@
+# Community and Contributing Guidelines for SynxFlow
+
+Welcome to the community for SynxFlow! We are excited to have you here and appreciate your interest in participating and contributing. To ensure a positive and productive experience for everyone, please read and adhere to the following guidelines.
+
+Our goal is to create a welcoming and inclusive community where everyone feels comfortable participating and contributing to SynxFlow, regardless of their background, experience level, or identity. This document outlines both our community standards and how you can contribute to the project.
+
+## Table of Contents
+
+* [Our Pledge (Code of Conduct)](#our-pledge-code-of-conduct)
+
+* [Contributing to SynxFlow](#contributing-to-synxflow)
+
+  * [Ways to Contribute](#ways-to-contribute)
+
+  * [Reporting Bugs](#reporting-bugs)
+
+  * [Suggesting Features](#suggesting-features)
+
+  * [Setting Up Your Development Environment](#setting-up-your-development-environment)
+
+  * [Making Code Contributions](#making-code-contributions)
+
+  * [Submitting a Pull Request](#submitting-a-pull-request)
+
+  * [Writing Documentation](#writing-documentation)
+
+
+* [Getting Help](#getting-help)
+
+* [Expectations](#expectations)
+
+* [Recognition](#recognition)
+
+* [License](#license)
+
+## Our Pledge (Code of Conduct)
+
+In the interest of fostering an open and welcoming environment, we as contributors and maintainers of SynxFlow pledge to making participation in our project and our community a harassment-free experience for everyone, regardless of age, body size, disability, ethnicity, sex characteristics, gender identity and expression, level of experience, education, socio-economic status, nationality, personal appearance, race, religion, or sexual identity and orientation.
+
+Examples of behavior that contributes to creating a positive environment include:
+
+* Using welcoming and inclusive language
+
+* Being respectful of differing viewpoints and experiences
+
+* Gracefully accepting constructive criticism
+
+* Focusing on what is best for the community
+
+* Showing empathy towards other community members
+
+Examples of unacceptable behavior by participants include:
+
+* The use of sexualized language or imagery and unwelcome sexual attention or advances
+
+* Trolling, insulting/derogatory comments, and personal or political attacks
+
+* Public or private harassment
+
+* Publishing others' private information, such as a physical or electronic address, without explicit permission
+
+* Other conduct which could reasonably be considered inappropriate in a professional setting
+
+For the full version of the Code of Conduct and details on enforcement, please refer to the [Contributor Covenant version 2.1](https://www.contributor-covenant.org/version/2/1/code_of_conduct.html).
+
+## Contributing to SynxFlow
+
+We welcome contributions of all kinds! Whether it's reporting a bug, suggesting a new feature, improving documentation, or submitting code, your contributions to SynxFlow are valuable.
+
+### Ways to Contribute
+
+There are many ways to contribute to SynxFlow beyond writing code:
+
+* **Reporting Bugs:** If you find a bug, please report it by opening an issue.
+
+* **Suggesting Features:** Have an idea for a new feature? Let us know by opening an issue.
+
+* **Improving Documentation:** Good documentation is crucial for SynxFlow. Help us improve it by submitting changes.
+
+* **Writing Tutorials or Blog Posts:** Share your knowledge of SynxFlow with others.
+
+* **Helping Other Users:** Participate in discussions and help answer questions about SynxFlow.
+
+* **Submitting Code:** Fix bugs, implement new features, or improve the existing codebase of SynxFlow.
+
+### Reporting Bugs
+
+If you encounter a bug in SynxFlow, please open a GitHub Issue. Before submitting, please search existing issues to see if the bug has already been reported.
+
[…]
