### 2025-05-12 #7586 @AnjaliSandip
@xiaxilin, could you fix the missing and invalid DOIs?
