### 2025-05-23 #7586 @AnjaliSandip
> [@xiaxilin](https://github.com/xiaxilin), could you fix the missing and invalid DOIs?

I am following up on this.  @xiaxilin, please complete the step above as soon as you can.
