### 2025-06-23 #7586 @xiaxilin
Hi @AnjaliSandip the latest version should be 1.0.2, regarding the DOI, should this be decided by the journal?
