### 2025-06-23 #7586 @AnjaliSandip
> Hi [@AnjaliSandip](https://github.com/AnjaliSandip) the latest version should be 1.0.2, regarding the DOI, should this be decided by the journal?

Please deposit a new archive of the software with Zenodo and share the DOI with me.
