### 2025-06-26 #7586 @kthyng
@xiaxilin Normally all submission authors are listed under "creators" on Zenodo so that the Zenodo archive is consistent with the JOSS submission. Can you make this update?
